move helper functions to its own module
